tests/performance/locust/requirements.txt
locust==2.24.0
PyJWT==2.8.0
